backend/requirements.txt
dash==2.10.2
dash_bootstrap_components==1.4.1
numpy==1.24.3
pandas==2.0.2
plotly==5.14.1
pymongo==4.3.3
python-dotenv==1.0.0
Requests==2.31.0
